--- lua/core/custom.lua
@@ -1,5 +1,0 @@
-local moduleExists = pcall(require, "custom")
-
-if moduleExists then
-   require "custom"
-end
