--- lua/core/init.lua
@@ -1,16 +1,0 @@
-local core_modules = {
-   "core.custom",
-   "core.options",
-   "core.autocmds",
-   "core.mappings",
-}
-
-for _, module in ipairs(core_modules) do
-   local ok, err = pcall(require, module)
-   if not ok then
-      error("Error loading " .. module .. "\n\n" .. err)
-   end
-end
-
--- set all the non plugin mappings
-require("core.mappings").misc()
